Add support for zip archives to import command
