+++ internal/cmd/importcmd.go
@@ -1,15 +1,7 @@
 package cmd
 
-// LATER add zip import using chezmoi.walkArchive
-
 import (
-	"archive/tar"
-	"bytes"
-	"compress/bzip2"
-	"compress/gzip"
-	"fmt"
 	"io"
-	"strings"
 
 	"github.com/spf13/cobra"
 
@@ -28,7 +20,7 @@ type importCmdConfig struct {
 func (c *Config) newImportCmd() *cobra.Command {
 	importCmd := &cobra.Command{
 		Use:     "import archive",
-		Short:   "Import a tar archive into the source state",
+		Short:   "Import an archive into the source state",
 		Long:    mustLongHelp("import"),
 		Example: example("import"),
 		Args:    cobra.MaximumNArgs(1),
@@ -52,33 +44,29 @@ func (c *Config) newImportCmd() *cobra.Command {
 }
 
 func (c *Config) runImportCmd(cmd *cobra.Command, args []string, sourceState *chezmoi.SourceState) error {
-	var r io.Reader
+	var (
+		name string
+		data []byte
+	)
 	if len(args) == 0 {
-		r = c.stdin
+		name = ".tar"
+		var err error
+		data, err = io.ReadAll(c.stdin)
+		if err != nil {
+			return err
+		}
 	} else {
 		absPath, err := chezmoi.NewAbsPathFromExtPath(args[0], c.homeDirAbsPath)
 		if err != nil {
 			return err
 		}
-		data, err := c.baseSystem.ReadFile(absPath)
+		name = absPath.String()
+		data, err = c.baseSystem.ReadFile(absPath)
 		if err != nil {
 			return err
 		}
-		r = bytes.NewReader(data)
-		switch base := strings.ToLower(absPath.Base()); {
-		case strings.HasSuffix(base, ".tar"):
-		case strings.HasSuffix(base, ".tar.bz2") || strings.HasSuffix(base, ".tbz2"):
-			r = bzip2.NewReader(r)
-		case strings.HasSuffix(base, ".tar.gz") || strings.HasSuffix(base, ".tgz"):
-			r, err = gzip.NewReader(r)
-			if err != nil {
-				return err
-			}
-		default:
-			return fmt.Errorf("unknown format: %s", base)
-		}
 	}
-	tarReaderSystem, err := chezmoi.NewTARReaderSystem(tar.NewReader(r), chezmoi.TARReaderSystemOptions{
+	archiveReaderSystem, err := chezmoi.NewArchiveReaderSystem(name, data, chezmoi.ArchiveReaderSystemOptions{
 		RootAbsPath:     c._import.destination,
 		StripComponents: c._import.stripComponents,
 	})
@@ -92,7 +80,7 @@ func (c *Config) runImportCmd(cmd *cobra.Command, args []string, sourceState *ch
 			return err
 		}
 	}
-	return sourceState.Add(c.sourceSystem, c.persistentState, tarReaderSystem, tarReaderSystem.FileInfos(), &chezmoi.AddOptions{
+	return sourceState.Add(c.sourceSystem, c.persistentState, archiveReaderSystem, archiveReaderSystem.FileInfos(), &chezmoi.AddOptions{
 		Exact:     c._import.exact,
 		Include:   c._import.include.Sub(c._import.exclude),
 		RemoveDir: removeDir,
